requirements.txt
apache-airflow==2.7.3 --constraint https://raw.githubusercontent.com/apache/airflow/constraints-2.7.3/constraints-3.10.txt
psycopg2-binary==2.9.4
ipykernel
pandas
sodapy
pyarrow
python-dotenv
confluent_kafka
fastavro
avro
streamlit
plotly
sqlalchemy
scikit-learn